[migrate-to-mdx] migrate quarterPicker
