+++ .storybook/stories/quarterPicker.stories.mdx
@@ -1,0 +1,35 @@
+import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks';
+import { withKnobs } from "@storybook/addon-knobs";
+import { withA11y } from '@storybook/addon-a11y';
+
+import QuarterPicker from '../../components/quarterPicker/QuarterPicker';
+import { QuarterPickerWrapper } from "../misc";
+
+<Meta
+    title="Quarter Picker"
+    component={QuarterPicker}
+    decorators={[withKnobs, withA11y]}
+    parameters={{
+        a11y: {
+            // ... axe options
+            element: '.story__container', // optional selector which element to inspect
+            config: {}, // axe-core configurationOptions (https://github.com/dequelabs/axe-core/blob/develop/doc/API.md#parameters-1)
+            options: {} // axe-core optionsParameter (https://github.com/dequelabs/axe-core/blob/develop/doc/API.md#options-parameter)
+        }
+    }}
+/>
+
+<Preview>
+    <Story name="Default">
+        <QuarterPickerWrapper>
+            <QuarterPicker disabledQuarters={[false, false, false, false]} />
+        </QuarterPickerWrapper>
+    </Story>
+</Preview>
+
+<Props of={QuarterPicker} />
+
+## Integration Status
+### USASpending
+- [ ] Download Center
+- [ ] Spending Explorer
