+++ .storybook/stories/tooltip.stories.mdx
@@ -38,8 +38,13 @@ import { Tooltip } from "../misc";
 
 <Props of={TooltipWrapper} />
 
-## USASpending Integration Status
+## Integration Status
+### USASpending
 - [x] Award Summary Page
-- [ ] Home Page, Hero
+- [ ] Home Page Hero
+- [ ] Advanced Search
+- [ ] Spending Explorer
+- [ ] State Profile Page
+- [ ] Agency Profile Page
 
-## Broker Integration Status
+### Broker Integration Status:
